split udisplay.cpp into multiple files
